* warnings-- for _core (signed/unsigned issues)
git-svn-id: file:///.data/svn_convert/svn@8318 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ branches/branch-playground/plugins/_core/msgview.cpp
@@ -1250,7 +1250,7 @@ void ViewParser::text(const QString &text)
     if (m_bInParagraph && (m_paragraphDir == DirAuto))
     {
         /* text isn't (unicode)-NULL terminated so we can't check for c->isNull! */
-        for(unsigned int i = 0; ((i < text.length()) && (m_paragraphDir == DirAuto)); i++)
+        for(unsigned int i = 0; ((i < (unsigned)text.length()) && (m_paragraphDir == DirAuto)); i++)
         {
             const QChar c = text.unicode()[i];
             // Note: Qt expects ltr/rtl to be lower-case.
